refactor(custom): Move SimpleDialog component into a shared component
